Split the `vector` module in submodules
